Clean PlotOptimals.py by exporting functions into ShapeFunctions.py
